Move LLMResponse object to its own file
